+++ src/connection.rs
@@ -12,7 +12,8 @@ use futures::{
     stream::LocalBoxStream,
 };
 use medea_client_api_proto::{
-    self as proto, ConnectionQualityScore, MemberId, TrackId,
+    self as proto, ConnectionQualityScore, MemberId, PeerConnectionState,
+    TrackId,
 };
 use tracerr::Traced;
 
@@ -254,6 +255,21 @@ impl Connections {
         self.members_to_conns.borrow().get(remote_member_id).cloned()
     }
 
+    /// Lists connections by [`TrackId`].
+    pub fn track_connections(
+        &self,
+        track_id: &TrackId,
+    ) -> Option<Vec<Connection>> {
+        self.tracks_to_members.borrow().get(track_id).map(|member_ids| {
+            member_ids
+                .iter()
+                .filter_map(|member_id| {
+                    self.members_to_conns.borrow().get(member_id).cloned()
+                })
+                .collect()
+        })
+    }
+
     /// Updates this [`Connection`] with the provided [`proto::state::Room`].
     pub fn apply(&self, new_state: &proto::state::Room) {
         #[expect(clippy::iter_over_hash_type, reason = "order doesn't matter")]
@@ -289,6 +305,49 @@ pub struct HandleDetachedError;
 #[derive(Clone, Debug)]
 pub struct ConnectionHandle(Weak<InnerConnection>);
 
+/// Estimated connection quality on client side.
+#[derive(Clone, Copy, Debug, Display, Eq, Ord, PartialEq, PartialOrd)]
+pub enum ClientConnectionQualityScore {
+    /// Connection is lost.
+    Disconnected = 0,
+
+    /// Nearly all users dissatisfied.
+    Poor = 1,
+
+    /// Many users dissatisfied.
+    Low = 2,
+
+    /// Some users dissatisfied.
+    Medium = 3,
+
+    /// Satisfied.
+    High = 4,
+}
+
+impl From<ConnectionQualityScore> for ClientConnectionQualityScore {
+    fn from(score: ConnectionQualityScore) -> Self {
+        match score {
+            ConnectionQualityScore::Poor => Self::Poor,
+            ConnectionQualityScore::Low => Self::Low,
+            ConnectionQualityScore::Medium => Self::Medium,
+            ConnectionQualityScore::High => Self::High,
+        }
+    }
+}
+
+/// [`Connection`]'s state.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum MemberConnectionState {
+    /// State in P2P mode.
+    P2P(PeerConnectionState),
+}
+
+impl From<PeerConnectionState> for MemberConnectionState {
+    fn from(state: PeerConnectionState) -> Self {
+        Self::P2P(state)
+    }
+}
+
 /// Actual data of a connection with a specific remote `Member`.
 ///
 /// Shared between external [`ConnectionHandle`] and Rust side [`Connection`].
@@ -299,6 +358,12 @@ struct InnerConnection {
 
     /// Current [`ConnectionQualityScore`] of this [`Connection`].
     quality_score: Cell<Option<ConnectionQualityScore>>,
+
+    /// Current [`ClientConnectionQualityScore`] of this [`Connection`].
+    client_quality_score: Cell<Option<ClientConnectionQualityScore>>,
+
+    /// Current [`MemberConnectionState`] of this [`Connection`].
+    state: Cell<Option<MemberConnectionState>>,
 
     /// Callback invoked when a [`remote::Track`] is received.
     on_remote_track_added: platform::Callback<api::RemoteMediaTrack>,
@@ -311,6 +376,9 @@ struct InnerConnection {
 
     /// Callback invoked when a [`ConnectionQualityScore`] is updated.
     on_quality_score_update: platform::Callback<u8>,
+
+    /// Callback invoked when a [`MemberConnectionState`] is updated.
+    on_state_change: platform::Callback<api::MemberConnectionState>,
 
     /// Callback invoked when this [`Connection`] is closed.
     on_close: platform::Callback<()>,
@@ -397,6 +465,36 @@ impl ConnectionHandle {
             .upgrade()
             .ok_or_else(|| tracerr::new!(HandleDetachedError))
             .map(|inner| inner.remote_id.0.clone())
+    }
+
+    /// Returns [`MemberConnectionState`] if it's known.
+    ///
+    /// # Errors
+    ///
+    /// See [`HandleDetachedError`] for details.
+    pub fn get_state(
+        &self,
+    ) -> Result<Option<MemberConnectionState>, Traced<HandleDetachedError>>
+    {
+        self.0
+            .upgrade()
+            .ok_or_else(|| tracerr::new!(HandleDetachedError))
+            .map(|inner| inner.state.get())
+    }
+
+    /// Sets callback, invoked when a new [`MemberConnectionState`] is set in this [`Connection`].
+    ///
+    /// # Errors
+    ///
+    /// See [`HandleDetachedError`] for details.
+    pub fn on_state_change(
+        &self,
+        f: platform::Function<api::MemberConnectionState>,
+    ) -> Result<(), Traced<HandleDetachedError>> {
+        self.0
+            .upgrade()
+            .ok_or_else(|| tracerr::new!(HandleDetachedError))
+            .map(|inner| inner.on_state_change.set_func(f))
     }
 
     /// Sets callback, invoked when a new [`remote::Track`] is added to this
@@ -579,7 +677,10 @@ impl Connection {
             ],
             remote_id,
             quality_score: Cell::default(),
+            client_quality_score: Cell::default(),
+            state: Cell::default(),
             on_quality_score_update: platform::Callback::default(),
+            on_state_change: platform::Callback::default(),
             recv_constraints,
             on_close: platform::Callback::default(),
             on_remote_track_added: platform::Callback::default(),
@@ -648,7 +749,47 @@ impl Connection {
 
     /// Updates [`ConnectionQualityScore`] of this [`Connection`].
     pub fn update_quality_score(&self, score: ConnectionQualityScore) {
-        if self.0.quality_score.replace(Some(score)) != Some(score) {
+        if self.0.quality_score.replace(Some(score)) == Some(score) {
+            return;
+        }
+
+        self.refresh_client_conn_quality_score();
+    }
+
+    /// Updates [`PeerConnectionState`] of this [`Connection`].
+    pub fn update_peer_state(&self, state: PeerConnectionState) {
+        let state = state.into();
+
+        if self.0.state.replace(Some(state)) == Some(state) {
+            return;
+        }
+
+        self.refresh_client_conn_quality_score();
+
+        self.0
+            .on_state_change
+            .call1::<api::MemberConnectionState>(state.into());
+    }
+
+    /// Refreshes [`ClientConnectionQualityScore`] of this [`Connection`].
+    fn refresh_client_conn_quality_score(&self) {
+        use MemberConnectionState as M;
+        use PeerConnectionState as S;
+
+        let state = self.0.state.get();
+        let quality_score = self.0.quality_score.get();
+        let score = match (state, quality_score) {
+            (Some(M::P2P(S::Connected)), Some(quality_score)) => quality_score.into(),
+            (Some(M::P2P(S::Disconnected | S::Failed | S::Closed)), _) => {
+                ClientConnectionQualityScore::Disconnected
+            }
+            (Some(M::P2P(S::Connecting | S::New)) | None, _)
+            | (Some(M::P2P(S::Connected)), None) => return,
+        };
+
+        let is_score_changed =
+            self.0.client_quality_score.replace(Some(score)) != Some(score);
+        if is_score_changed {
             // TODO: Replace with derive?
             #[expect(clippy::as_conversions, reason = "needs refactoring")]
             self.0.on_quality_score_update.call1(score as u8);