+++ tests/web.rs
@@ -77,6 +77,7 @@ macro_rules! js_callback {
     }}
 }
 
+mod connection;
 mod media;
 mod peer;
 mod room;